refactor bbol event store
